app/requirements.txt
flask
vanna[milvus,openai,mysql]
db-dtypes
python-dotenv
cryptography>=45.0.5